Create pre-requirements.sh
This notebook relies on an old version of `hstcal` and therefore an old version of `calwf3`. Every `hstcal` version > 2.5.0 and every corresponding `calwf3` version > 3.5.2 will employ the v2.0 PCTE correction. Since this notebook's purpose is to demonstrate how to use the v1.0 PCTE correction, we must ensure `hstcal` version 2.5.0 is installed and used. I don't believe this version (2.5.0) is offered on conda-forge.