Revised DrizzlePac Notebook: align_to_catalogs (#265)
* Revised align_to_catalogs.ipynb

This notebook has been revised in consultation with J. Mack and the DrizzlePac team. The updates include an improved workflow for GAIA alignments that solve previously encountered errors in the notebook, as well as numerous STScI DMD formatting and PEP8 compliance changes. The notebook was tested extensively with stenv2024.02.05.

* Revised requirements.txt

The revised requirements.txt is identical to the previous file in terms of required packages, with updated version numbers that were used in testing with stenv2024.02.05.

* Revised align_to_catalogs.ipynb

Added links to the text that point users to critical sections of the DrizzlePac Handbook.

* Revised requirements.txt

The DMD continuous integration checks were failing because the CRDS package, which is included in stenv, was not explicitly listed in the requirements.txt file. The following packages were added to the requirements.txt file: crds, hdf5 (which is required to pip install drizzlepac), python, and jupyter.

* Revised requirements.txt

The DMD continuous integration checks failed when including hdf5==1.14.3, which was needed to pip install drizzlepac in an empty environment, and has now been removed.

* Revised requirements.txt

The DMD continuous integration checks failed when including python==3.11.7, which was needed to initialize an empty environment, and has now been removed.

* Incorporated DMD Suggestions

This commit incorporates the suggestions based on the DMD review by Hatice Karatay. Specifically, the STScI logo was removed from the header, there is now only one main level in the markdown, all figures have titles and axes labels, and several code blocks were revised to include file handling checks. See pull request #265 for additional details.