Revised DrizzlePac Notebook: using_updated_astrometry_solutions (#275)
* Revised using_updated_astrometry_solutions.ipynb

This notebook has been revised in consultation with J. Mack and the DrizzlePac team. The updates include an improved workflow for WCS solutions that solve previously encountered errors in the notebook, as well as numerous STScI DMD formatting and PEP8 compliance changes. The notebook was tested extensively with stenv2024.02.05.

* Revised requirements.txt

The revised requirements.txt is identical to the previous file in terms of required packages, with updated version numbers that were used in testing with stenv2024.02.05.

* Cleared cell outputs

Cleared the output messages from all cells.

* updated documentation

The documentation for several code cells and the notebooks overall workflow were improved by Jennifer Mack.

* Apply context manager for file handling

---------

Co-authored-by: Hatice Karatay <haticekaraty@gmail.com>