Revised sky matching notebook (#284)
* Delete notebooks/DrizzlePac/sky_matching/MDRIZSKY_Values.png

this figure is no longer needed in the new version of the notebook.  We instead make the figure manually

* Delete notebooks/DrizzlePac/sky_matching/drz.list

The file is no longer needed in the new version of the notebook. Instead we manually generate the list of DRZs to be passed to the functions

* Update sky_matching.ipynb

updated workflow and markdown text

* Update requirements.txt

the notebook no longer depends on ccdproc or stwcs

* Update requirements.txt

pinning photutils to see if it fixes the error in the CI

* Update sky_matching.ipynb

made changes suggested by Hatice

* Update requirements.txt

* Remove deprecated show_in_notebook method

---------

Co-authored-by: Hatice Karatay <haticekaraty@gmail.com>